Move list_contains helper to Utils
